Experimenting with several algorithms
Implemented a trie, whose best performance with tries and hash tables
of ids of tries was comparable to previous solution.

Experimented with tries of the whole word, and other stuff...

Finally divided the triplets in a list of lists containing words of the same
size, so that hashes are computed from those words with same lenght.
For each of those hashes, we get the winning triplet and then we
merge with previous winning triplet. That way we end up with the
highest winning triplet.

I think this algorithm improves performance because the number of
collissions is lowered.

Code will be cleaned up on next commit.