Adding a couple of optimizations to reduce problem size
First optimization consists in converting the input alphabet
to '0' to 'Z' (in uppercase), that way math simplifies to do histograms
and memory is kept low. Instead of spaces  colon ':' is used internally.

The other optimization is to iterate all strings of same size using
an histogram and remove many words so the hash table takes less time.